+++ src/Control/Monad/Bayes/Sampler.hs
@@ -27,6 +27,7 @@ module Control.Monad.Bayes.Sampler
     sampleSTfixed,
     toBins,
     sampleMean,
+    sampler,
   )
 where
 
@@ -70,6 +71,9 @@ newtype SamplerIO a = SamplerIO (ReaderT GenIO IO a)
 -- the operating system.
 -- For efficiency this operation should be applied at the very end, ideally
 -- once per program.
+sampler :: SamplerIO a -> IO a
+sampler = sampleIO
+
 sampleIO :: SamplerIO a -> IO a
 sampleIO (SamplerIO m) = createSystemRandom >>= runReaderT m